refactor: split action core modules
